Modified Setting model to make it polymorphic Modified settings controller to allow for new model
To allow user and domain settings of different types to be stored easily the Setting model was modified to make it polymorphic. The base Setting model is now abstract and only used for searching. There are seven new Setting child models for storing different types of data.
 - StringSetting
 - DateSetting
 - DateTimeSetting
 - IntegerSetting
 - FloatSetting
 - BlobSetting
 - BooleanSetting
The settings controller was also modified to take the new types into account when creating or retrieving setting data.