Modified settings_for_identity to return all settings with affect the identity
Function now returns all the settings which impact the user. This is done by first finding all domain settings and then layering the user settings on top. It is important to remember that User Settings > Domain Settings. For situations where the user should not be allowed to overwrite a setting enforce this via the UI.